requirements.txt
Package       Version
------------- -------
discord.py    2.4.0
pip           24.2
python-dotenv 1.0.1
requests      2.32.3
setuptools    58.0.4
